feat: Add username/password entry
This commit added the "login_username_entry" and "login_password_entry" elements to the fragment_login.xml, and added getText() functionality for those textboxes in LoginFragment.java. Additionally, the login screen's UI was redesigned a bit to show the app's name, and to space textboxes and buttons appropriately.

Additionally, two Toast messages displaying the entered Username and Password were prepared, for testing.

*Updated UI: fragment_login.xml
*Added functionality: LoginFragment.java
*Testing functionality: LoginFragment.java lines 50 to 59

Addresses #10