feat: User/Password query
This commit added to LoginFragment.java. Now, upon clicking the button labeled "Login (To Home)" a query is sent to the Firestore database's "users" collection, searching for a document with "username" and "password" fields whose values match those entered by the user.

Toast messages display, conveying whether the entered username and password are valid. If they are valid, the button progresses to the Home screen fragment. If at least one value is not valid, or a textbox is left empty, the user stays on the Login screen fragment.

*Exceptions: Logcat TAG and error written if retrival of document fails.
*Database functionality: LoginFragment.java connects to database, and sends query
*User Experience: Toast messages read "Logging in" or "Invalid Username or Password"

Addresses #10