feat: User/Password saved locally
This commit now saves user data (Username and Password Strings) to the default SharedPreferences file for this app. This also sets the groundwork for any functions requiring the use of SharedPreferences in the future.

The use of SharedPreferences allows for variable values (or other Objects) to be shared across classes within the app. This will be the way a user's username will be accessed for the sake of personalization, and for generating Firestore queries regarding user information. (Ex. Logging an exercise to "exercises_[USERNAME]" by calling getUsername() to retrieve that [USERNAME] from SharedPreferences.

Toast messages display after a successful login for the sake of testing this commit. These Toast messages display the user's username and password retrieved from SharedPreferences, showcasing the setUsernameAndPassword() function works as intended.

Files changed:
*LoginFragment.java
*MainActivity.java

*Storage: SharedPreferences accesses the default preferences file for this app, only this app, and is inaccessible from other apps.
*Testing: Username and Password are displayed using Toast messages upon login. This is temporary, and will be removed upon next commit.
*Future benefit: The functions that access and alter SharedPreferences are written in MainActivity.java, the constant activity across all Fragments. Thus, these functions are accessible from throughout the app, allowing for the true implementation of a user-login system across all applicable features.
*Coding base: The specific code needed to call these functions are noted above each function in MainActivity.java.

Addresses #14