Feat: Logging an Exercise PBI
Update: Added functionality to validate user input data correctly via the isInputValid method. All fields have to be filled out.
Updated the Hints on the page so the user knows what text is expected in each field. Removed Activity because we will be handling persistence in a different way.
The gradle is updated to handle Java, Kotlin and Anything else.
Log exercise Grabs all text from the fields and stores them in a map object
Then writes to the DB in the exercise collection.

* added logExercise Method
* added isInputValid Method
* added correct Hints in the .xml
* changed functionality so that you can log multiple exercises
* Changed button color for discard

Addresses issue #13

addresses #13