feat: New user registration
This commit adds quite a few small features, practically completely fleshing out the greater feature of user registration.

First, fragment_register.xml has been redesigned to house four textboxes for user entry of Username, First Name, Email, and Password, as well as a new "Back to login" button that returns the user to the login screen without attempting to register the information in the textboxes.
Second, AndroidManifest.xml has a permission added to it to allow for access to the host device's network connectivity state.
Third, and biggest, RegisterFragment.java has been updated with the following changes:
* The "isOnline" function returns true if the host device is connected to a network. Please note that this does NOT confirm internet connectivity, meaning connecting to a public network but not signing in leads to a "true" return statement. This is addressed later.
* The "registration" function takes four Strings as parameters: the four Strings entered by the user. This function first checks if the current database has a matching Username or Email string in its "users" collection. If no match is found, a new document is created housing the entered user data, and the user is taken to the Login screen signifying a successful user registration.

The "Register new user" button in this fragment now does the following when clicked, in order:
* Obtain the Strings entered in all four textboxes.
* Confirm the entered email contains @ character
* Confirm the entered password is 8 or more characters in length
* Confirm all textboxes are not empty
* Call ifOnline() function to check for device's network connection
* If connected, attempt to read a specific collection from the Firestore database, specifying SERVER as the source of the query. This query is used to confirm that the device is able to connect to the serverside database. If this test is not done, then the user could theoretically register and login as a new user, while only using the locally cached database. This solution took hours to brainstorm, and is necessary.
* Only if this query from the server is successful, the "registration" function is called, passing all four user-entered Strings as parameters. At this point (within registration function) the database is checked for a user with existing username/email.

This took much longer than anticipated, but I am happy with the results quite a bit. The reason a test query to the server's database needed to take place before registration is attempted is because, when testing on a physical device, I discovered that connecting to a public network without signing in did not prevent database write-queries, which could have caused catastrophic results if not addressed in the first push.

Toast messages display at all possible avenues to inform of registration status:
* Network connection is offline
* Connection to serverside database fails
* Email does not contain @ symbol
* Password is too short
* Textbox is left empty
* Registration is successful

Files changed:
*RegisterFragment.java
*fragment_register.xml
*AndroidManifest.xml

Addresses #22