fix: add scrollView for table and change retrieving collection from Firestore
This commit added a scrollView element for the exercise table log so it will be able to show enough exercises once they excide the table length. It also changed the collection name that the function uses from Firestore and the variable names to make it compatiable with other issues.

* added a scrollView elements in fragment_prev_activity.xml
* changed the collections name in fillTable() in PrevActivityFragment.java to "exercises"
* changed the field names in fillTable() in PrevActivityFragment.java to match the new "exercises" collection

issue #12