Update: Added Functionality for Deny
* Added Listeners to the Buttons
* Included Functionality for Deny Request

Noticed an issue with the display and removal of requests for a user. A user can have multiple requests from one person. Not a problem when looping through the documents but when it comes to the removal of those requests the only way i can think of to remove them is to nest a for loop in a for loop. which can get crunchy.

Think an easier way to handle this would be to address Issue #11 and not allow for a user to send a request to someone they've already sent a request to.

Addresses Issue #34