Update: Functionality For Adding a Friend
* The On-Click Listener is Fleshed Out. Adds a Requesting User's Username to the Friends List associated with the User. Removes the row from the table. Removes the request document.

Same thing goes. This is bare-bones functionality. No testing of ... Does the user already exist in the friends collection. I think a better way of handling this is to make sure that the requestingUser cant request to be friends with someone they're already friends with.
Issue #11.
A thought ... if someone is denied being friends they shouldnt be able to continue to harass the user by constantly sending friend requests.

Addresses Issue #34