feat: Hashing entered passwords
This commit creates a function in MainActivity.java called "hashString" that takes a String as parameter, and hashes (encrypts) it using the SHA-512 method. This function is called in LoginFragment.java and RegisterFragment.java to hash entered passwords before they are used in queries to the Firestore database.

This commit (currently) causes all previously registered users to be incapable of logging in due to their passwords in the database being raw unencrypted Strings.

Additionally, a comment was removed from HomeFragment.java causing a false error to be reported in the IDE.

Files changed:
 * HomeFragment.java
 * LoginFragment.java
 * RegisterFragment.java
 * MainActivity.java

Addresses #29