feat: Staying logged in
This commit adds two functions to MainActivity.java that serve the purpose of retaining a user's login status. The "checkForUsername()" function returns true if a username is stored in SharedPreferences. This is called in the onCreate() function for LoginFragment.java, and if the result is true, then the app automatically naviagates directly to the Home Fragment, instead of displaying the Login Fragment. The other function added to MainActivity.java is "clearSharedPref()", and serves to remove the "username" and "password" keys in the SharedPreferences file. This is called in ProfileFragment.java when the "Logout" button is clicked.

Additionally, some small changes were made for the sake of functionality:
* Removed a comment from ProfileFragment.java that was being flagged as an error.
* Changed all instances of "getPreferences" function calls in MainActivity.java to "getSharedPreferences" function calls, in order to specify the name of the SharedPreferences file being accessed.
* Updated the "setUsernameAndPassword" function in MainActivity.java so that the user's entered password is NOT stored in SharedPreferences, for the sake of security. A DUMMY password is stored instead.

Files changed:
* LoginFragment.java
* ProfileFragment.java
* MainActivity.java

Addresses #30