test: Improvements to previous tests
Previously written tests rely on spending time logging in before running a test function. Since this PBI removes the login screen after the device has previously logged in to the app, this causes most tests to fail. As such, a function has been added to every previously utilized test file that clears the SharedPreferences file, which is called before and after every test.

Additionally, the "blankFieldsTest()" from LogExercise_IT.java was updated to test for blank fields in a manner that doesn't rely on synchronous testing environment, since the previous method of reading the text from a displayed Toast message was consistently failing on multiple machines across the team.

Files changed:
* Friends_IT.java
* LogExercise_IT.java
* LogIn_IT.java
* PreviousActivity_IT.java

Addresses #30