feat: Inbox UI and Functionality
The InboxFragment now shows a table layout containing all documents from the "inbox_[USERNAME]" collection in FireStore, where [USERNAME] is the user currently logged in. This is done using an altered version of the fillTable() function used in PrevActivity.

The inbox documents are pulled from FireStore ordered by Date, and every TableRow placed within the TableLayout has an onClick listener assigned to it. This onClick listener displays an AlertDialog box that prompts the user to select from four encouraging responses, and has "No, thanks" and "Send" button options. After "Send" is clicked, the selected reply is sent in a document to a FireStore sub-collection with the following path:
"exercises_[USERNAME OF ORIGINAL USER]"
->"[ORIGINAL EXERCISE DOCUMENT]"
->"replies"
->"[GENERATED NAME OF REPLY]".

There is currently no functionality that allows this user to view these replies in Resolute. When a reply is sent the document is deleted from "inbox_[USERNAME]", and the InboxFragment is recreated to reflect this deletion.

Files changed:
* nav_graph.xml
* InboxFragment.java
* fragment_inbox.xml

Addresses #15