Reply to All Function:
This commit adds the new Reply All Function where a new button is added to the fragment_inbox.xml file and when the button is clicked a pop up dialouge appears and and when a message is clicked and sent the function sends it to all the rows and deletes it from the table and the firestore.Ive made many changes for this to work and to also have the orginal method to still be functional. I will be commiting all the new changes. Issue #58