feat: Table fills with replies
The TableLayout on ReplyFragment now fills with replies from the "replies" sub-collection of an appropriate exercise. The fillTable() function in ReplyFragment.java is very similar to other fillTable() functions in InboxFragment.java or PrevActivityFragment.java. Replies are sorted by most recent.

Additionally, I changed the colors of the TableLayout header and the back button in fragment_reply.xml.

Files changed:
* ReplyFragment.java
* fragment_reply.xml

Addresses #45