test: ReplyFragment related testing
The ReplyFragment was updated to now include text stating "No replies yet..." if a logged exercise has no replies sub-collection.

There are now tests for ReplyFragment functionality in the file Reply_IT.java. The tests check that an exercise with no replies displays the "No replies yet..." text, and that two different exercises show their respective contents of their own replies sub-collections, which is basically a test checking that the TableLayout used to show replies empties and fills correctly when navigating to ReplyFragment.

A linear test was added that was needed for Inbox_IT.java. It tests that an exercise in a user's inbox can be replied to, and that the correct reply is viewable from that original user's ReplyFragment for the respective exercise. This is linear until exercise deletion is implemented, and incomplete until sending exercises to inboxes is merged.

Also, some unused import statements were removed from PreviousActivity_IT.java.

Files changed:
* ReplyFragment.java
* Inbox_IT.java
* PreviousActivity_IT.java

Files added:
* Reply_IT.java

Addresses #45