fix: Navigation bug involving tables
The patched bug saw the app crash everytime navigation between pages took place too quickly, when one of those pages contained a table. This is because if the app navigated away from a table before the query finished that filled the table with data from FireStore, the app would then try to fill a TableLayout that does not exist on screen.

This bug was fixed by moving all navigation buttons' onClick() listeners to a separate function. This function is only called when one of three conditions is met: the query succeeds and the table fills, the query fails and the table does not fill, or the user's device is offline and cannot gather data from FireStore. Online connectivity is checked with a new function "isOnline()". There are also toast messages informing the user that their device is offline.

These checks take place on all pages involving a table, except for PrevActivityFragment. On this screen, the "isOnline()" function is not implemented. As such, the user can still view their previous activity while offline, using cached database collections. Navigation will still only be possible after this table fills.

Files changed:
* ReplyFragment.java
* FriendsFragment.java
* InboxFragment.java
* PrevActivityFragment.java
* RequestFragment.java

Addresses #57